+++ CHANGELOG.rst
@@ -4,6 +4,14 @@ Change Log
 
 This document records all notable changes to **u2py**.
 
+**1.0** (2017-08-10)
+-------------------------
+* Phase change to Beta
+* Added support for PyQt4 version
+* Removed installation time check for PyQt5 version
+* Added runtime check for PyQt4 and PyQt5
+* Added configuration option to install PyQt4 or PyQt5
+* Changed project structure - added 'converters' pacakage
 
 **0.2a0** (2017-07-24)
 -------------------------